requirements.txt
openai
wikipedia
duckduckgo_search
arxiv
requests
python-dotenv
sreamlit